// src/definitions/enums/delegate.enum.ts
export enum TabNames {
  CHOXULY = "waitHandleTab",
  DAXULY = "handledTab",
}

export enum SearchTitles {
  UPDATEDATE = "UPDATE_DATE",
  CREATEDATE = "CREATE_DATE",
  NUMBERSIGN = "NUMBER_SIGN",
  PREVIEW = "PREVIEW",
  USER_ENTER = "USER_ENTER",
  DOCTYPE = "DOC_TYPE",
  STATUS = "STATUS",
  PERSON_HANDLE = "HANDLE_USER",
  SECURITY_NAME = "SECURITY",
  HANDLE_DATE = "HANDLE_DATE",
  DELEGATE_USER = "DELEGATE_USER",
}

export enum DelegateTabNames {
  MAIN_HANDLE = "MAIN_HANDLE",
  COMBINE_HANDLE = "COMBINE_HANDLE",
  HANDLED = "HANDLED",
}

export enum DelegateSearchTitles {
  UPDATEDATE = "UPDATEDATE",
  CREATEDATE = "CREATEDATE",
  NUMBER_ARRIVAL = "NUMBER_ARRIVAL",
  DATE_ARRIVAL = "DATE_ARRIVAL",
  DATE_RECEIVED = "DATE_RECEIVED",
  DEADLINE = "DEADLINE",
  DOCID = "DOCID",
  NUMBERSIGN = "NUMBERSIGN",
  PREVIEW = "PREVIEW",
  ORG_ISSUED_NAME = "ORG_ISSUED_NAME",
  ORG_HANDLE = "ORG_HANDLE",
  DOCTYPE = "DOCTYPE",
  DATEISSUED = "DATEISSUED",
  PERSON_HANDLE = "PERSON_HANDLE",
  SECURITY_NAME = "SECURITY_NAME",
  STATUS = "STATUS",
  DELEGATER = "DELEGATER",
  PROGRESS = "PROGRESS",
  PROCESS_STATUS = "PROCESS_STATUS",
}